requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
tqdm
Pillow
pacmap
opencv-python
tlc
